remove inmemory db
