+++ go.mod
@@ -1,10 +1,3 @@
 module go-to-school/main
 
 go 1.21.4
-
-require (
-	github.com/google/uuid v1.4.0
-	github.com/hashicorp/go-immutable-radix v1.3.0 // indirect
-	github.com/hashicorp/go-memdb v1.3.4 // indirect
-	github.com/hashicorp/golang-lru v0.5.4 // indirect
-)
